Fixed pedantic warnings in windows

+++ src/SceneGame.cpp
@@ -793,12 +793,12 @@ bool SceneGame::DoCrate( Crate& crate )
 				// Only add points first time switch is visited
 				if ( crate.connected->type == Object::CONVEYOR
 					and ((Conveyor*)(crate.connected))->isSwitch == true
-					and crate.visited.find( level.GetIndex( crate.connected->pos.x, crate.connected->pos.y ) ) == crate.visited.end() )
+					and crate.visited.find( level.GetIndex( (int)crate.connected->pos.x, (int)crate.connected->pos.y ) ) == crate.visited.end() )
 				{
 					crate.value += 25;
 
 					// Mark this switch so it doesn't get counted again
-					crate.visited.insert( level.GetIndex( crate.connected->pos.x, crate.connected->pos.y ) );
+					crate.visited.insert( level.GetIndex( (int)crate.connected->pos.x, (int)crate.connected->pos.y ) );
 				}
 			}
